Add files via upload
Question 1:
Implement this game as one Python function. The function should have one parameter which is used to select whether a Unicorn or a Pegasus is built. Follow the steps and logic you take in a physical game. Your code should include enough print statements to generate a clear nontrivial example illustrating a play of the game, telling the story as the dice is rolled and the creature’s body parts are assembled. Explain carefully what steps you have taken to test your code (not just debugging). Please provide both your code and the output from one sample run for each of Unicorn and Pegasus.

Implemented the dice game as one python function. The function has one parameter which is used to select whether a unicorn or Pegasus is being built.
Below code implements the game. If the user enters unicorn/Pegasus as input then the code runs the function in such a way that it collects all the body parts of the unicorn/Pegasus.
If the user enters any other string as input other than unicorn or Pegasus then he should get the message “invalid game selection” and the program will be ended.