Create adding counts for the unicorn and pegasus game and also playing Unicorn and pegasus games seperately
Task 2:
Question 2:
Modify your Python code to add at least three counts that effectively summaries a game, e.g., 
how many dice rolls are made. Use these summary counts to compare building a Unicorn with building a Pegasus. 
Also implement a variant in which the player builds one complete Unicorn and one complete Pegasus in parallel (while carefully 
satisfying the precedence of body parts for both creatures), and compare the resulting counts with those of the original game (making some relevant conclusions).
Please provide separate Python code for the original game with counts and the variant, together with output from sample runs of each and a comparison of counts in 
a table for the Unicorn-only, Pegasus-only and both-in-parallel versions.

	Added 3 counts in the code that effectively summaries a game. 
	Count -- Total number of rolls to collect all the body parts.
	Count  -- Total number of times we get wings while we roll the dice through out the game.
	Count --  Total number of times we get horn while we roll the dice through out the game.
I have counted the number of times a player gets horn and wing because it is one of the key deciding element to complete the game.