Create Code for playing both in parallel versions:
	Comparing the resulting counts with playing Unicorn game or Pegasus game or both-in parallel versions.
Results	Unicorn only	Pegasus only	Both in parallel
Test Run 1	number of rolls to collect all the body parts of unicorn :  27

number of times we get wings while we roll the dice through out the game: 3

number of times we get horn while we roll the dice through out the game: 1
	number of rolls to collect all the body_parts of pegasus :  67

number of times we get wings while we roll the dice through out the game: 12

number of times we get horn while we roll the dice through out the game: 7
	Unicorn	Pegasus
number of rolls to collect all the body_parts of unicorn :  27

number of times we get wings while we roll the dice through out the game: 2

number of times we get horn while we roll the dice through out the game: 3
	number of rolls to collect all the body_parts of pegasus :  44

number of times we get wings while we roll the dice through out the game: 4

number of times we get horn while we roll the dice through out the game: 8


Test Run 2	number of rolls to collect all the body parts of unicorn :  29

number of times we get wings while we roll the dice through out the game: 4

number of times we get horn while we roll the dice through out the game: 3
	number of rolls to collect all the body parts of pegasus :  57

number of times we get wings while we roll the dice through out the game: 4

number of times we get horn while we roll the dice through out the game: 6
	Pegasus	Unicorn
number of rolls to collect all the body_parts of pegasus :  54

number of times we get wings while we roll the dice through out the game: 3

number of times we get horn while we roll the dice through out the game: 4
	number of rolls to collect all the body_parts of unicorn :  52

number of times we get wings while we roll the dice through out the game: 9

number of times we get horn while we roll the dice through out the game: 2


Test Run 3	number of rolls to collect all the body parts of unicorn :  52

number of times we get wings while we roll the dice through out the game: 6

number of times we get horn while we roll the dice through out the game: 6	number of rolls to collect all the body parts of Pegasus :  77

number of times we get wings while we roll the dice through out the game: 5

number of times we get horn while we roll the dice through out the game: 16	Pegasus	Unicorn
number of rolls to collect all the body_parts of pegasus :  40

number of times we get wings while we roll the dice through out the game: 6

number of times we get horn while we roll the dice through out the game: 6
	number of rolls to collect all the body_parts of unicorn :  64

number of times we get wings while we roll the dice through out the game: 10

number of times we get horn while we roll the dice through out the game: 7


Conclusion	For playing unicorn, we need to collect 1 horn. We have more chance of completing the game with less number of rolls when compared with Pegasus game.	For playing Pegasus, we need to collect 2wings. We need more chances of rolling dice for completing the game  when compared with unicorn game.	 Pegasus 	Unicorn
multiprocessing time taken for pegasus : 0.0005197525024414062	multiprocessing time taken for unicorn : 0.0007100105285644531
multiprocessing time taken for pegasus : 0.0008435249328613281	multiprocessing time taken for unicorn : 0.0009143352508544922

multiprocessing time taken for pegasus : 0.00047087669372558594
	multiprocessing time taken for unicorn : 0.0003769397735595703
The time taken to run both the versions in parallel is nearly equal